NBLAST: add mgmt. command to generate cluster tasks
This can be used to create cluster tasks that compute NBLAST based on
dividing the query object space in parts of equal aggregated length.
E.g. if the total length of 10 skeletons is 500nm (each 50nm)
and we want to compute an all by all NBLAST result with five parallel
tasks (cluster nodes), this management command would generate five shell
scripts that would compute two skeletons versus all ten each. Results
are stored in a relational fashion.

The script also allows to generate boilerplate code for creating a local
SSH tunnel and activating a conda or venv environment.

This could be a more complex run with a local SSH tunnel in each task to
a remote Postgres using key based authentication:

./manage.py catmaid_parallel_nblast --similarity_id 40 --create-tasks --target-dir=jobs --n-jobs=100 --ssh true --ssh-local-port 7777 --ssh-remote-port 5432 --ssh-host tom@spine.int.janelia.org --ssh-identity /home/user/.ssh/id_rsa --conda catmaid