Add new connector type for mitochondria
A mitochondria_of relation can now be used to link treenodes to
connector nodes. It is available through the <Alt>+Click menu to select
connector types. The linked mitochondria skeletons can be seen in the
Connectivity Widget, the 3D Viewer and the Morphology plot. They are
currently rendered in a purple color.

Thanks to Anton Miroschnikow for this suggestion.